backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==2.0.6
cliapp==1.20180812.1
click==8.0.3
cloudinary==1.42.2
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.3.0
dj-rest-auth==7.0.0
Django==5.1
django-admin-interface==0.29.4
django-allauth==65.2.0
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cloudinary-storage==0.3.0
django-colorfield==0.12.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-grappelli==4.0.1
django-js-asset==2.2.0
django-taggit==6.0.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.15.4
frozenlist==1.2.0
future==0.18.2
GDAL==3.4.1
gunicorn==23.0.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
multidict==5.1.0
mysql-connector-python==9.0.0
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
protobuf==3.12.4
proton-core==0.4.0
proton-keyring-linux==0.2.0
proton-vpn-api-core==0.42.4
proton-vpn-gtk-app==4.9.6
proton-vpn-network-manager==0.12.13
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.19.2
python-gnupg==0.4.8
python-slugify==8.0.4
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.5.1
systemd-python==234
text-unidecode==1.3
texttable==1.6.4
ttystatus==0.38
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.3
wadllib==1.3.6
websocket-client==1.2.3
whitenoise==6.9.0
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
